feat: working on seperate the connect window
